Support for strata in the resampling.
Idea was to use with JGL, but it turns out to be a bit redundant.

JGL has two regularization parameters, and the recommendation is to
tune lambda1 first. However this is the same (almost) as tuning
lambda on a standard glasso with the entire dataset.

You'd then use a second layer of tuning to select lambda2.

Thus the changes I've made aren't very interesting.